rename all single choice to finite quiz
