+++ .nginx/compose/nginx-dev.conf
@@ -18,7 +18,7 @@ server {
         proxy_pass      http://overworld-backend/api/;
     }
 
-    location /minigames/singlechoice/api/ {
+    location /minigames/finitequiz/api/ {
         proxy_pass      http://host.docker.internal:8080/api/;
     }
 
@@ -26,8 +26,8 @@ server {
         proxy_pass      http://overworld/;
     }
 
-#     location /minigames/singlechoice/ {
-#         proxy_pass      http://singlechoice/;
+#     location /minigames/finitequiz/ {
+#         proxy_pass      http://finitequiz/;
 #     }
 
     location /lecturer-interface/ {
