+++ .nginx/compose/nginx.conf
@@ -18,16 +18,16 @@ server {
         proxy_pass      http://overworld-backend/api/;
     }
 
-    location /minigames/singlechoice/api/ {
-        proxy_pass      http://Singlechoice-backend/api/;
+    location /minigames/finitequiz/api/ {
+        proxy_pass      http://finitequiz-backend/api/;
     }
 
     location /overworld/ {
         proxy_pass      http://overworld/;
     }
 
-    location /minigames/singlechoice/ {
-        proxy_pass      http://singlechoice/;
+    location /minigames/finitequiz/ {
+        proxy_pass      http://finitequiz/;
     }
 
     location /lecturer-interface/ {
