+++ Dockerfile
@@ -12,9 +12,9 @@ RUN mvn -f /home/app/pom.xml package -Dmaven.test.skip
 # Package stage
 #
 FROM openjdk:17-jdk-slim
-COPY --from=build /home/app/target/singlechoice-backend-0.0.1-SNAPSHOT.jar /usr/local/lib/singlechoice-backend.jar
+COPY --from=build /home/app/target/finitequiz-backend-0.0.1-SNAPSHOT.jar /usr/local/lib/finitequiz-backend.jar
 EXPOSE 80
 ENV POSTGRES_URL "postgresql://localhost:5432/postgres"
 ENV POSTGRES_USER "postgres"
 ENV POSTGRES_PASSWORD "postgres"
-ENTRYPOINT /usr/local/openjdk-17/bin/java -jar /usr/local/lib/singlechoice-backend.jar --spring.datasource.url=jdbc:${POSTGRES_URL} --server.port=80 --spring.datasource.username=${POSTGRES_USER} --spring.datasource.password=${POSTGRES_PASSWORD}
+ENTRYPOINT /usr/local/openjdk-17/bin/java -jar /usr/local/lib/finitequiz-backend.jar --spring.datasource.url=jdbc:${POSTGRES_URL} --server.port=80 --spring.datasource.username=${POSTGRES_USER} --spring.datasource.password=${POSTGRES_PASSWORD}
