+++ docker-compose-dev.yaml
@@ -12,7 +12,7 @@ services:
       POSTGRES_SETUP: >
         login/default.sql
         overworld/default.sql
-        chickenshock/default.sql
+        finitequiz/default.sql
 
   login-db:
     container_name: login-db
@@ -34,8 +34,8 @@ services:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
 
-  singlechoice-db:
-    container_name: chickenshock-db
+  finitequiz-db:
+    container_name: finitequiz-db
     image: postgres:14-alpine
     restart: always
     ports:
@@ -88,9 +88,9 @@ services:
     expose:
       - "80"
 
-#  singlechoice:
-#    container_name: singlechoice
-#    image: ghcr.io/gamify-it/singlechoice:main
+#  finitequiz:
+#    container_name: finitequiz
+#    image: ghcr.io/gamify-it/finitequiz:main
 #    restart: always
 #    pull_policy: always
 #    expose:
