+++ docker-compose.yaml
@@ -5,14 +5,14 @@ services:
   #databases
   setup:
     container_name: setup
-    image: ghcr.io/gamify-it/test-data:feature-singlechoice-data-dump-2
+    image: ghcr.io/gamify-it/test-data:feature-finitequiz-data-dump-2
     pull_policy: always
     restart: on-failure
     environment:
       POSTGRES_SETUP: >
         login/default.sql
         overworld/default.sql
-        singlechoice/default.sql
+        finitequiz/default.sql
 
   login-db:
     container_name: login-db
@@ -34,8 +34,8 @@ services:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
 
-  singlechoice-db:
-    container_name: singlechoice-db
+  finitequiz-db:
+    container_name: finitequiz-db
     image: postgres:14-alpine
     restart: always
     ports:
@@ -71,18 +71,18 @@ services:
       - POSTGRES_PASSWORD=postgres
       - POSTGRES_URL=postgresql://overworld-db:5432/postgres
 
-  singlechoice-backend:
-    container_name: singlechoice-backend
+  finitequiz-backend:
+    container_name: finitequiz-backend
     build: .
     restart: always
     expose:
       - "80"
     depends_on:
-      - singlechoice-db
+      - finitequiz-db
     environment:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
-      - POSTGRES_URL=postgresql://singlechoice-db:5432/postgres
+      - POSTGRES_URL=postgresql://finitequiz-db:5432/postgres
 
 
   #frontends
@@ -102,9 +102,9 @@ services:
     expose:
       - "80"
 
-  singlechoice:
-    container_name: singlechoice
-    image: ghcr.io/gamify-it/singlechoice:main
+  finitequiz:
+    container_name: finitequiz
+    image: ghcr.io/gamify-it/finitequiz:main
     restart: always
     pull_policy: always
     expose:
