+++ sonar-project.properties
@@ -1,1 +1,1 @@
-sonar.projectKey=singlechoice-backend
+sonar.projectKey=finitequiz-backend
