+++ src/main/java/de/unistuttgart/finitequizbackend/FinitequizServiceApplication.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.singlechoicebackend;
+package de.unistuttgart.finitequizbackend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,9 +6,9 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
 @EnableFeignClients
-public class SinglechoiceServiceApplication {
+public class FinitequizServiceApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(SinglechoiceServiceApplication.class, args);
+    SpringApplication.run(FinitequizServiceApplication.class, args);
   }
 }
