+++ src/main/java/de/unistuttgart/finitequizbackend/data/ConfigurationDTO.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.singlechoicebackend.data;
+package de.unistuttgart.finitequizbackend.data;
 
 import java.util.Objects;
 import java.util.Set;
