+++ src/main/java/de/unistuttgart/finitequizbackend/data/GameResult.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.singlechoicebackend.data;
+package de.unistuttgart.finitequizbackend.data;
 
 import java.time.LocalDateTime;
 import java.util.List;
