+++ src/main/java/de/unistuttgart/finitequizbackend/repositories/GameResultRepository.java
@@ -1,6 +1,6 @@
-package de.unistuttgart.singlechoicebackend.repositories;
+package de.unistuttgart.finitequizbackend.repositories;
 
-import de.unistuttgart.singlechoicebackend.data.GameResult;
+import de.unistuttgart.finitequizbackend.data.GameResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
