+++ src/main/java/de/unistuttgart/finitequizbackend/service/AuthorizationService.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.singlechoicebackend.service;
+package de.unistuttgart.finitequizbackend.service;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
